Maybe OP
Added: crossbow_otty, arrow_tnt, arrow_stone;
Modified: - init.lua, arrow.lua.
crossbow_otty : a new weapon put in init.lua. This weapon with a silly name (I'm Marco Ottina) is a indestructible bow which cast the right-near arrow if present, or a "normal" arrow if no arrow is found, consuming it its count is greater than 1 or it's not a fire arrow. If the count is exactly 1, the arrow is gratis.
tnt_arrow : a simple arrow that explode like a TNT, but with a radius of 2. Burn the tile it touch
stone_arrow : a cheaper arrow than Arrow, but deals 2 damage instead of 3
Arrow : added new cheaper recipes, but they gives less arrows